Move non-template registry implementation out of public header
`include/zoo/tools/registry.hpp` carried ~400 lines of non-template
implementation (schema normalization, JSON Schema validation helpers,
registry storage operations) that were compiled into every translation
unit that included the header. None of those functions required template
instantiation.

Move them to a new `src/tools/registry.cpp`. Keep template-dependent
callable traits, `make_tool_definition<Func>`, and the three typed
`register_tool` overloads inline. All public names and signatures are
preserved — no consumer source changes required.

Header: 724 → 382 lines. New TU: 409 lines.
